Rename the package to impact-stack-rest
This is to accomodate pylint not working with implicit namespace
packages.

See https://github.com/pylint-dev/pylint/issues/7875 or one of its many
duplicates.

isort’s auto-detection seems to need some extra config to recognize the package.
